// ssyweb/SSYNEW/src/app/models/fundReleaseOrder.model.ts
export class FundReleaseOrder {
    fundReleaseOrderHdrId: number;
    rloOfficeID: number;
    rloOfficeName: string;
    requestedBy: string;
    fundRequestNumber: number;
    requestDate: string;
    requestDateString: string;
    amount: number;
    sanctionedAmount: number;
    statusId: number;
    statusName: string;
    createdBy: number;
    selected: boolean;
    boardName: any;
    boardId: any;
    alcId: number;
    fundReleaseDate: string;
    fundReleaseDateString: string;
}